fix make_whole() for double precision boxes

+++ package/MDAnalysis/lib/_cutil.pyx
@@ -216,7 +216,8 @@ def make_whole(atomgroup, reference_atom=None, inplace=True):
             raise ValueError("Reference atom not in atomgroup")
         ref = ix_to_rel[reference_atom.ix]
 
-    box = atomgroup.dimensions
+    box = atomgroup.dimensions.astype(np.float32)
+    # TODO: remove astype(np.float32) once all universes return float32 boxes
 
     for i in range(3):
         half_box[i] = 0.5 * box[i]
